Added notch settings sliders

+++ MusicNotch/Variables.swift
@@ -15,7 +15,7 @@ extension Defaults.Keys {
     
     static let launchAtLogin = Key<Bool>("LaunchAtLogin", default: LaunchAtLogin.isEnabled)
     
-    static let hideNotchTime = Key<Int>("hideNotchTime", default: 3)
+    static let hideNotchTime = Key<CGFloat>("hideNotchTime", default: 3)
     static let notchSizeChange = Key<CGFloat>("notchSizeChange", default: 0)
 }
